Migrate to platform APIs
Reapplies #47
