+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2024 ENLYZE GmbH <hello@enlyze.com>
+Copyright 2025 ENLYZE GmbH <hello@enlyze.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of this
 software and associated documentation files (the "Software"), to deal in the Software
