+++ README.rst
@@ -65,8 +65,8 @@ so:*
 
     $ export PYENCHANT_LIBRARY_PATH=/opt/homebrew/lib/libenchant-2.dylib
 
-
 Examples
---------------------------------
-You can find examples of how to use the Python SDK here: `Notebooks <examples/notebooks>`_
+--------
 
+You can find examples of how to use the Python SDK here: `Notebooks
+<examples/notebooks>`_
