+++ docs/api_client/client.rst
@@ -1,0 +1,7 @@
+Platform API Client
+===================
+
+.. currentmodule:: enlyze.api_client.client
+
+.. autoclass:: PlatformApiClient()
+    :members:
