+++ docs/api_client/models.rst
@@ -1,17 +1,35 @@
 Models
 ======
 
-.. currentmodule:: enlyze.api_clients.production_runs.models
+.. currentmodule:: enlyze.api_client.models
 
-.. autoclass:: ProductionRunsApiModel()
+.. autoclass:: PlatformApiModel()
 
-.. autoclass:: ProductionRun()
+.. autoclass:: Site()
     :members:
     :undoc-members:
     :exclude-members: model_config, model_fields
     :show-inheritance:
 
 .. autoclass:: Machine()
+    :members:
+    :undoc-members:
+    :exclude-members: model_config, model_fields
+    :show-inheritance:
+
+.. autoclass:: Variable()
+    :members:
+    :undoc-members:
+    :exclude-members: model_config, model_fields
+    :show-inheritance:
+
+.. autoclass:: TimeseriesData()
+    :members:
+    :undoc-members:
+    :exclude-members: model_config, model_fields
+    :show-inheritance:
+
+.. autoclass:: ProductionRun()
     :members:
     :undoc-members:
     :exclude-members: model_config, model_fields
