--- docs/api_clients/base.rst
@@ -1,17 +1,0 @@
-Base Client
-===========
-
-.. currentmodule:: enlyze.api_clients.base
-
-.. autoclass:: M
-
-.. autoclass:: R
-
-.. autoclass:: ApiBaseModel
-
-.. autoclass:: PaginatedResponseBaseModel
-
-.. autoclass:: ApiBaseClient
-    :members:
-    :private-members:
-    :undoc-members:
