--- docs/api_clients/index.rst
@@ -1,9 +1,0 @@
-API Clients
-===========
-
-.. toctree::
-    :maxdepth: 1
-
-    base
-    timeseries/index
-    production_runs/index
