--- docs/api_clients/production_runs/client.rst
@@ -1,9 +1,0 @@
-Production Runs API Client
-==========================
-
-.. currentmodule:: enlyze.api_clients.production_runs.client
-
-.. autoclass:: _PaginatedResponse
-
-.. autoclass:: ProductionRunsApiClient()
-    :members:
