--- docs/api_clients/production_runs/index.rst
@@ -1,8 +1,0 @@
-Production Runs API
-===================
-
-.. toctree::
-    :maxdepth: 1
-
-    client
-    models
