--- docs/api_clients/timeseries/client.rst
@@ -1,11 +1,0 @@
-Timeseries API Client
-=====================
-
-.. currentmodule:: enlyze.api_clients.timeseries.client
-
-.. autoclass:: _PaginatedResponse()
-    :members:
-    :exclude-members: model_config, model_fields
-
-.. autoclass:: TimeseriesApiClient()
-    :members: get, get_paginated
