--- docs/api_clients/timeseries/models.rst
@@ -1,30 +1,0 @@
-Models
-======
-
-.. currentmodule:: enlyze.api_clients.timeseries.models
-
-.. autoclass:: TimeseriesApiModel()
-
-.. autoclass:: Site()
-    :members:
-    :undoc-members:
-    :exclude-members: model_config, model_fields
-    :show-inheritance:
-
-.. autoclass:: Machine()
-    :members:
-    :undoc-members:
-    :exclude-members: model_config, model_fields
-    :show-inheritance:
-
-.. autoclass:: Variable()
-    :members:
-    :undoc-members:
-    :exclude-members: model_config, model_fields
-    :show-inheritance:
-
-.. autoclass:: TimeseriesData()
-    :members:
-    :undoc-members:
-    :exclude-members: model_config, model_fields
-    :show-inheritance:
