+++ docs/examples.rst
@@ -1,7 +1,10 @@
 Examples
 ========
-In our examples section, you can find Jupyter Notebooks and other material to help you get started:
+
+In our examples section, you can find Jupyter Notebooks and other material to help you
+get started:
 
 `Notebooks <https://github.com/enlyze/enlyze-python/tree/master/examples/notebooks>`_:
 
-* `Introduction to the ENLYZE Python SDK <https://github.com/enlyze/enlyze-python/tree/master/examples/notebooks/introduction_enlyze_python_sdk.ipynb>`_
+- `Introduction to the ENLYZE Python SDK
+  <https://github.com/enlyze/enlyze-python/tree/master/examples/notebooks/introduction_enlyze_python_sdk.ipynb>`_
